Isolate tests that explicitly break auto-loading guarantees.
